# food-retail-marketplace/ardo-backend/docs/swagger.yaml
definitions:
...
  management_user.AddUserIn:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      preferredLang:
        type: string
      userType:
        type: string
    type: object
  management_user.UpdateUserCredentialsIn:
    properties:
      firstName:
        type: string
      lastName:
        type: string
    type: object
  management_user.UpdateUserEmailIn:
    properties:
      email:
        type: string
    type: object
  management_user.UpdateUserPasswordIn:
    properties:
      password:
        type: string
    type: object
  management_user.UpdateUserPreferredLangIn:
    properties:
      lang:
        type: string
    type: object
  management_user.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isDeleted:
        type: boolean
      lastName:
        type: string
      preferredLang:
        type: string
      updatedAt:
        type: string
      userType:
        type: string
      version:
        type: integer
    type: object
  management_user.UsersData:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/management_user.User'
        type: array
    type: object
  response.Error:
    properties:
      messages:
        example:
        - error message
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
      traceId:
        example: string
        type: string
    type: object
  response.Success:
    properties:
      data: {}
      success:
        example: true
        type: boolean
      traceId:
        example: string
        type: string
    type: object
host: api.ardogroup.org
info:
  contact:
    name: API Support
    url: https://nassyrov.t.me
  description: Detailed info about all endpoints
  title: Ardo Backend API
  version: 1.0.0
paths:
...
  /management/users:
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Is deleted
        in: query
        name: is_deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_user.UsersData'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get users by filters
      tags:
      - Management Users
    post:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Add user object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_user.AddUserIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Add user
      tags:
      - Management Users
  /management/users/{user_id}:
    delete:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Delete user
      tags:
      - Management Users
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_user.User'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get user by id
      tags:
      - Management Users
  /management/users/credentials/{user_id}:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update user credentials object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_user.UpdateUserCredentialsIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update user credentials
      tags:
      - Management Users
  /management/users/email/{user_id}:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update user email object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_user.UpdateUserEmailIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update user email
      tags:
      - Management Users
  /management/users/password/{user_id}:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update user password object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_user.UpdateUserPasswordIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update user preferred lang
      tags:
      - Management Users
  /management/users/preferred-lang/{user_id}:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update user preferred lang object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_user.UpdateUserPreferredLangIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update user preferred lang
      tags:
      - Management Users
  /management/users/recover/{user_id}:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Recover user
      tags:
      - Management Users
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"